Fixed compile warning.

+++ src/7.in_practice/2.text_rendering/text_rendering.cpp
@@ -145,7 +145,7 @@ int main()
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x
+                static_cast<unsigned int>(face->glyph->advance.x)
             };
             Characters.insert(std::pair<char, Character>(c, character));
         }
@@ -258,4 +258,3 @@ void RenderText(Shader &shader, std::string text, float x, float y, float scale,
     glBindVertexArray(0);
     glBindTexture(GL_TEXTURE_2D, 0);
 }
-
